Move workflows index query operations into the manager layer.
